discord_bot/requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
certifi==2022.6.15
chardet==4.0.0
charset-normalizer==2.1.0
discord==1.7.3
discord.ext==0.8
discord.py==1.7.3
idna==3.3
mongoengine==0.24.1
multidict==6.0.2
praw==7.6.0
prawcore==2.3.0
pymongo==4.1.1
python-dotenv==0.20.0
requests==2.28.1
typing_extensions==4.3.0
update-checker==0.18.0
urllib3==1.26.9
websocket-client==1.3.3
yarl==1.7.2
